renamed folder for pages and deleted unused database folder
